Restructure to allow for BaseStorage interface
