+++ schematic/models/metadata.py
@@ -12,12 +12,12 @@ from typing import Any, Dict, Optional, Text, List
 from schematic.schemas.explorer import SchemaExplorer
 from schematic.manifest.generator import ManifestGenerator
 from schematic.schemas.generator import SchemaGenerator
-from schematic.synapse.store import SynapseStorage
+from schematic.store.synapse import SynapseStorage
 from schematic.utils.df_utils import trim_commas_df
 
 class MetadataModel(object):
     """Metadata model wrapper around schema.org specification graph.
-    
+
     Provides basic utilities to:
 
     1) manipulate the metadata model
@@ -64,18 +64,18 @@ class MetadataModel(object):
     # controller components are (loosely speaking) responsible for handling the interaction between views and the model
     # some of these components right now reside in the Bundle class
 
-    def getModelSubgraph(self, rootNode: str, 
+    def getModelSubgraph(self, rootNode: str,
                         subgraphType: str) -> nx.DiGraph:
         """Gets a schema subgraph from rootNode descendants based on edge/node properties of type subgraphType.
-        
+
         Args:
             rootNode: a schema node label (i.e. term).
             subgraphType: the kind of subgraph to traverse (i.e. based on node properties or edge labels).
-        
+
         Returns:
             A directed subgraph (networkx DiGraph) of the metadata model with vertex set root node descendants.
 
-        Raises: 
+        Raises:
             ValueError: rootNode not found in metadata model.
         """
         pass
@@ -86,11 +86,11 @@ class MetadataModel(object):
         Args:
             rootNode: a schema object/node label (i.e. term)
             relationshipType: edge label type of the schema subgraph (e.g. requiresDependency)
-        
+
         Returns:
             An ordered list of objects, that are all descendants of rootNode.
 
-        Raises: 
+        Raises:
             ValueError: rootNode not found in metadata model.
         """
         ordered_nodes = self.sg.get_descendants_by_edge_type(rootNode, relationshipType, connected=True, ordered=True)
@@ -99,19 +99,19 @@ class MetadataModel(object):
 
         return ordered_nodes
 
-    
-    def getModelManifest(self, title: str, rootNode: str, jsonSchema: str = None, filenames: list = None) -> str: 
+
+    def getModelManifest(self, title: str, rootNode: str, jsonSchema: str = None, filenames: list = None) -> str:
         """Gets data from the annotations manifest file.
 
         TBD: Does this method belong here or in manifest generator?
-        
+
         Args:
             rootNode: a schema node label (i.e. term).
-        
+
         Returns:
             A manifest URI (assume Google doc for now).
 
-        Raises: 
+        Raises:
             ValueError: rootNode not found in metadata model.
         """
         additionalMetadata = {}
@@ -137,64 +137,64 @@ class MetadataModel(object):
         """Given a source model component (see https://w3id.org/biolink/vocab/category for definnition of component), return all components required by it.
         Useful to construct requirement dependencies not only between specific attributes but also between categories/components of attributes;
         Can be utilized to track metadata completion progress across multiple categories of attributes.
-        
-        Args: 
+
+        Args:
             source_component: an attribute label indicating the source component.
 
-        Returns: 
+        Returns:
             A list of required components associated with the source component.
-        """ 
+        """
         # get metadata model schema graph
         # mm_graph = self.se.get_nx_schema()
 
         # get required components for the input/source component
         req_components = self.sg.get_component_requirements(source_component)
-        # req_components = get_component_requirements(mm_graph, source_component) 
+        # req_components = get_component_requirements(mm_graph, source_component)
 
         return req_components
 
 
     # TODO: abstract validation in its own module
-    def validateModelManifest(self, manifestPath: str, rootNode: str, jsonSchema: str = None) -> List[str]:     
+    def validateModelManifest(self, manifestPath: str, rootNode: str, jsonSchema: str = None) -> List[str]:
         """Check if provided annotations manifest dataframe satisfies all model requirements.
 
         Args:
             rootNode: a schema node label (i.e. term).
             manifestPath: a path to the manifest csv file containing annotations.
-        
+
         Returns:
             A validation status message; if there is an error the message.
             contains the manifest annotation record (i.e. row) that is invalid, along with the validation error associated with this record.
-        
-        Raises: 
+
+        Raises:
             ValueError: rootNode not found in metadata model.
         """
         # get validation schema for a given node in the data model, if the user has not provided input validation schema
         if not jsonSchema:
             jsonSchema = self.sg.get_json_schema_requirements(rootNode, rootNode + "_validation")
-         
+
         errors = []
- 
+
         # get annotations from manifest (array of json annotations corresponding to manifest rows)
         manifest = pd.read_csv(manifestPath)    # read manifest csv file as is from manifest path
         manifest = trim_commas_df(manifest).fillna("")  # apply cleaning logic as part of pre-processing step
- 
+
         # check if each of the provided annotation columns has validation rule 'list'
         # if so, assume annotation for this column are comma separated list of multi-value annotations
         # convert multi-valued annotations to list
         for col in manifest.columns:
-            
+
             # remove trailing/leading whitespaces from manifest
             manifest.applymap(lambda x: x.strip() if isinstance(x, str) else x)
 
             # convert manifest values to string
-            # TODO: when validation handles annotation types as validation rules 
+            # TODO: when validation handles annotation types as validation rules
             # would have to avoid converting everything to string
             manifest[col] = manifest[col].astype(str)
 
             # if the validation rule is set to list, convert items in the
             # annotations manifest to a list and strip each value from leading/trailing spaces
-            if "list" in self.sg.get_node_validation_rules(col): 
+            if "list" in self.sg.get_node_validation_rules(col):
                 manifest[col] = manifest[col].apply(lambda x: [s.strip() for s in str(x).split(",")])
 
         annotations = json.loads(manifest.to_json(orient='records'))
@@ -203,37 +203,37 @@ class MetadataModel(object):
 
             for error in sorted(v.iter_errors(annotation), key=exceptions.relevance):
                 errorRow = i + 2
-                errorCol = error.path[-1] if len(error.path) > 0 else "Wrong schema" 
+                errorCol = error.path[-1] if len(error.path) > 0 else "Wrong schema"
                 errorMsg = error.message[0:500]
                 errorVal = error.instance if len(error.path) > 0 else "Wrong schema"
 
                 errors.append([errorRow, errorCol, errorMsg, errorVal])
-                
+
         return errors
 
 
-    def populateModelManifest(self, title, manifestPath: str, rootNode: str) -> str: 
+    def populateModelManifest(self, title, manifestPath: str, rootNode: str) -> str:
         """Populate an existing annotations manifest based on a dataframe.
-         
+
         Args:
             rootNode: a schema node label (i.e. term).
             manifestPath: a path to the manifest csv file containing annotations.
-        
+
         Returns:
             A link to the filled in model manifest (e.g. google sheet).
 
-        Raises: 
+        Raises:
             ValueError: rootNode not found in metadata model.
         """
         mg = ManifestGenerator(title, self.inputMModelLocation, rootNode)
-        
+
         emptyManifestURL = mg.get_manifest()
 
         return mg.populate_manifest_spreadsheet(manifestPath, emptyManifestURL)
 
 
     def submit_metadata_manifest(self, manifest_path: str, dataset_id: str, validate_component: str = None) -> bool:
-        """Wrap methods that are responsible for validation of manifests for a given component, and association of the 
+        """Wrap methods that are responsible for validation of manifests for a given component, and association of the
         same manifest file with a specified dataset.
         Args:
             manifest_path: Path to the manifest file, which contains the metadata.
@@ -265,8 +265,8 @@ class MetadataModel(object):
 
             # if there are no errors in validation process
             if not val_errors:
-                
-                # upload manifest file from `manifest_path` path to entity with Syn ID `dataset_id` 
+
+                # upload manifest file from `manifest_path` path to entity with Syn ID `dataset_id`
                 syn_store.associateMetadataWithFiles(metadataManifestPath=manifest_path, datasetId=dataset_id)
 
                 print("No validation errors resulted during association.")
@@ -279,5 +279,5 @@ class MetadataModel(object):
         syn_store.associateMetadataWithFiles(metadataManifestPath=manifest_path, datasetId=dataset_id)
 
         print("Validation was not performed on manifest file before association.")
-        
+
         return True
