+++ schematic/store/__init__.py
@@ -1,0 +1,1 @@
+from schematic.store.synapse import SynapseStorage
