+++ schematic/store/base.py
@@ -1,0 +1,6 @@
+class BaseStorage(object):
+    """Common interface for data store back-ends (like Synapse and Gen3)."""
+
+    def __init__(self):
+        """Base function for initializing object."""
+        raise NotImplementedError()
