--- schematic/synapse/__init__.py
@@ -1,1 +1,0 @@
-from schematic.synapse.store import SynapseStorage
